+++ src/rules/plugins/typescript.ts
@@ -7,7 +7,9 @@ export = {
   "consistent-return": "off",
   "@typescript-eslint/consistent-return": "error",
   "@typescript-eslint/consistent-type-exports": "error",
-  "@typescript-eslint/consistent-type-imports": "error",
+  "@typescript-eslint/consistent-type-imports": [ "error", {
+    prefer: "type-imports"
+  }],
   "default-param-last": "off",
   "@typescript-eslint/default-param-last": "error",
   "@typescript-eslint/explicit-member-accessibility": "error",