+++ src/rules/plugins/unused-imports.ts
@@ -1,7 +1,0 @@
-import type { Linter } from "eslint";
-
-// https://github.com/sweepline/eslint-plugin-unused-imports
-export = {
-  "unused-imports/no-unused-imports": "error",
-  "unused-imports/no-unused-vars": "error"
-} satisfies Linter.Config["rules"];